[Auto Generated] 1.1.2

+++ src/main/java/com/sdk/platform/models/order/OrderModels.kt
@@ -5198,19 +5198,19 @@ data class FinancialBreakup(
     
     
     @SerializedName("refund_credit")
-    var refundCredit: Int?=null,
+    var refundCredit: Double?=null,
     
     @SerializedName("amount_paid_roundoff")
     var amountPaidRoundoff: Int?=null,
     
     @SerializedName("price_effective")
-    var priceEffective: Int?=null,
+    var priceEffective: Double?=null,
     
     @SerializedName("promotion_effective_discount")
     var promotionEffectiveDiscount: Double?=null,
     
     @SerializedName("transfer_price")
-    var transferPrice: Int?=null,
+    var transferPrice: Double?=null,
     
     @SerializedName("coupon_effective_discount")
     var couponEffectiveDiscount: Double?=null,
@@ -5219,7 +5219,7 @@ data class FinancialBreakup(
     var gstFee: Double?=null,
     
     @SerializedName("tax_collected_at_source")
-    var taxCollectedAtSource: Int?=null,
+    var taxCollectedAtSource: Double?=null,
     
     @SerializedName("brand_calculated_amount")
     var brandCalculatedAmount: Double?=null,
@@ -5234,7 +5234,7 @@ data class FinancialBreakup(
     var hsnCode: String?=null,
     
     @SerializedName("cashback")
-    var cashback: Int?=null,
+    var cashback: Double?=null,
     
     @SerializedName("item_name")
     var itemName: String?=null,
@@ -5243,13 +5243,13 @@ data class FinancialBreakup(
     var valueOfGood: Double?=null,
     
     @SerializedName("cashback_applied")
-    var cashbackApplied: Int?=null,
+    var cashbackApplied: Double?=null,
     
     @SerializedName("cod_charges")
-    var codCharges: Int?=null,
+    var codCharges: Double?=null,
     
     @SerializedName("price_marked")
-    var priceMarked: Int?=null,
+    var priceMarked: Double?=null,
     
     @SerializedName("size")
     var size: String?=null,
@@ -5261,13 +5261,13 @@ data class FinancialBreakup(
     var couponValue: Double?=null,
     
     @SerializedName("discount")
-    var discount: Int?=null,
+    var discount: Double?=null,
     
     @SerializedName("fynd_credits")
-    var fyndCredits: Int?=null,
+    var fyndCredits: Double?=null,
     
     @SerializedName("gst_tax_percentage")
-    var gstTaxPercentage: Int?=null,
+    var gstTaxPercentage: Double?=null,
     
     @SerializedName("identifiers")
     var identifiers: Identifier?=null,
